# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ILO DCMS API",
        "description": "API documentation for ILO Dispute Case Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/case-natures": {
            "get": {
                "tags": [
                    "Case Natures"
                ],
                "summary": "Get case natures",
                "description": "Returns a paginated list of case natures with optional search and status filtering.",
                "operationId": "43f08cb91e2bb40ab3d36e2b564428f3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by case nature name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "By Government"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by active/inactive status.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of records per page. Maximum 100.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case natures retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Case Natures"
                ],
                "summary": "Create case nature",
                "description": "Create a new case nature.",
                "operationId": "699dbe49da0110c4cecca4d7d288489f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_en"
                                ],
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "By Government",
                                        "maxLength": 255
                                    },
                                    "name_bn": {
                                        "type": "string",
                                        "example": "সরকার কর্তৃক",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true,
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Case nature created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/case-natures/{caseNature}": {
            "get": {
                "tags": [
                    "Case Natures"
                ],
                "summary": "Get case nature",
                "description": "Returns a single case nature by ID.",
                "operationId": "7d00d58593ad5e65efaa96ec9d8b664d",
                "parameters": [
                    {
                        "name": "caseNature",
                        "in": "path",
                        "description": "Case nature ID.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case nature retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Case nature not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Case Natures"
                ],
                "summary": "Update case nature",
                "description": "Update an existing case nature.",
                "operationId": "3bea32e3f3f5e9b49148e626bd544319",
                "parameters": [
                    {
                        "name": "caseNature",
                        "in": "path",
                        "description": "Case nature ID.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "By Government",
                                        "maxLength": 255
                                    },
                                    "name_bn": {
                                        "type": "string",
                                        "example": "সরকার কর্তৃক",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Case nature updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Case nature not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Case Natures"
                ],
                "summary": "Delete case nature",
                "description": "Delete an existing case nature.",
                "operationId": "ea56658ff7eb65ad1f3878f4b351fb3b",
                "parameters": [
                    {
                        "name": "caseNature",
                        "in": "path",
                        "description": "Case nature ID.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case nature deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Case nature not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}